+++ src/main/java/fr/estecka/variantscit/modules/CustomDataModule.java
@@ -11,20 +11,23 @@ import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
 import net.minecraft.util.Identifier;
 
-public class CustomVariantModule
+public class CustomDataModule
 implements ISimpleCitModule
 {
-	static public final MapCodec<String> CODEC = RecordCodecBuilder.mapCodec(builder->builder
+	static public final MapCodec<CustomDataModule> CODEC = RecordCodecBuilder.mapCodec(builder->builder
 		.group(
-			Codec.STRING.fieldOf("nbtKey").forGetter(s->s)
+			Codec.STRING.fieldOf("nbtKey").forGetter(s->s.key),
+			Codec.BOOL.fieldOf("caseSensitive").orElse(true).forGetter(s->s.caseSensitive)
 		)
-		.apply(builder, s->s)
+		.apply(builder, CustomDataModule::new)
 	);
 
 	private final String key;
+	private final boolean caseSensitive;
 
-	public CustomVariantModule(String key){
+	public CustomDataModule(String key, boolean caseSensitive){
 		this.key = key;
+		this.caseSensitive = caseSensitive;
 	}
 
 	@Override
@@ -35,6 +38,10 @@ implements ISimpleCitModule
 		if (component==null || (nbt=component.getNbt())==null || !nbt.contains(key, NbtElement.STRING_TYPE))
 			return null;
 
-		return Identifier.tryParse(nbt.getString(key));
+		String rawVariant = nbt.getString(key);
+		if (!caseSensitive)
+			rawVariant = rawVariant.toLowerCase();
+
+		return Identifier.tryParse(rawVariant);
 	}
 }